iter35 cluster-037: Mainnet Responses/Messages host 改走 Application command facade
错误模式: ResponsesEndpoints/MessagesEndpoints 注入大量 application/runtime collaborators + Host 内执行 caller resolution / chat-route decision / session registration / model route parsing / tool classification / LLMRequest construction / execution / SSE shaping。

新原则: Host handlers 仅 parse/auth HTTP + 委托 typed Application command facade(owns Normalize → Resolve Target → Build Context → Dispatch/Observe)。SSE rendering 留 boundary。

⟦AI:AUTO-LOOP⟧
